move constants into constants file. add more types to ui elements strategy
